+++ src/client/java/legends/ultra/cool/addons/hud/widget/Bar.java
@@ -6,10 +6,10 @@ import net.minecraft.client.gui.DrawContext;
 
 public class Bar extends HudWidget {
     private final String ownerId;
-    private int min;
-    private int max;
+    private double min;
+    private double max;
 
-    public Bar(String ownerId, String name, double x, double y, int min, int max) {
+    public Bar(String ownerId, String name, double x, double y, double min, double max) {
         super(name, x, y);
         this.ownerId = ownerId;
         this.min = min;
@@ -22,11 +22,11 @@ public class Bar extends HudWidget {
             return;
         }
 
-        float width = getWidth();
-        float height = getHeight();
-        float healthWidth = 0;
+        double width = getWidth();
+        double height = getHeight();
+        double healthWidth = 0;
         if (max > 0) {
-            healthWidth = width * min / (float) max;
+            healthWidth = width * (float)min / (float) max;
         }
 
         context.fill((int) (x - 3), (int) (y - 3), (int) (x + width + 2), (int) (y + height + 2), 0x50000000);
@@ -35,22 +35,22 @@ public class Bar extends HudWidget {
     }
 
     @Override
-    public int getWidth() {
+    public double getWidth() {
         return WidgetConfigManager.getInt(ownerId, "barWidth", 80);
     }
 
     @Override
-    public int getHeight() {
+    public double getHeight() {
         return 3;
     }
 
-    public void setMin(int min) {
+    public void setMin(double min) {
         if (min >= 0) {
             this.min = min;
         }
     }
 
-    public void setMax(int max) {
+    public void setMax(double max) {
         if (max > 0) {
             this.max = max;
         }
